+++ app/preprocessing/bible/thessalonians1/__init__.py
@@ -1,0 +1,3 @@
+"""
+Init for 1 Thessalonians.
+"""